Update patch set 2
Patch Set 2:

> As we discussed on irc and as you highlight in your comment, we
 > need to have a common way for handling the scenarios. To me whether
 > a scenarios is generic or specific does not matter at this phase so
 > we need common way of handling all the scenarios.
 > 
 > My thinking when I proposed this change and tried to summarize irc
 > is based on how osa advices people to create new roles. If we apply
 > that to OPNFV scenarios, I end up with treating scenarios as roles,
 > thus proposing this patch.
 > 
 > When it comes to how the roles for generic scenarios get connected
 > to XCI; same as the roles for the specific scenarios which you
 > proposed in change; https://gerrit.opnfv.org/gerrit/#/c/44881/ and
 > put in to file opnfv-scenario-requirements.yml.

Can you give an example of how an entry in opnfv-scenario-requirements.yml would look like for a role that already lives in the XCI repository? That file was aimed to be used by external projects to put their info there so we know how to clone/configure/link them to the XCI prior deployment. Generic roles live in the XCI already, so we have prior knowledge of what needs to be done to get them properly linked. As such, external and generic roles differ on that regard. We can simply symlink xci/scenario/foobar -> xci/playbooks/roles when we create a new scenario. Would that work for you?

Patch-set: 2