Update patch set 2
Patch Set 2:

Here is the example listing these scenarios 

https://paste.ubuntu.com/25747821/

Again, this is not really about 
- us knowing the scenarios or 
- whether they are physically located in xci repo or somewhere else or
- (currently) owned and maintained by xci team. 

They are OPNFV scenarios and should be treated/handled in same way. 

Symlinking doesn't work either as you do things differently for the scenarios depending on where they live.

These will not be the latest generic scenarios we will have and I really think we should try to keep modifications to xci itself as little as possible once we settle the basics, meaning that scenarios stay kind of external to xci. So we basically expect each and every scenario to go through same path.

Another point is that it will be difficult to explain how the scenarios handled if we have different way of cloning/configuring the generic and specific scenarios to people. With this simple alignment and perhaps some extra job on our side, we can keep things aligned right at the very beginning, making it easier for us to explain things and people to get the idea.

Another point is; the generic scenarios currently live in releng-xci repo but this does not mean they will be there in future so if they are moved to somewhere else, we will hopefully not need to do anything apart from updating src and role path. (see the new project proposal clover which might take over k8 scenario...)

Patch-set: 2
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0