Update patch set 1
Patch Set 1: Code-Review-2

You really shouldn't touch the start-new-vm.sh since we use it for the CI and you likely don't have to use it for your deployment (at least not yet).

If we add proxy support it should be done for all distros and all flavors not just AIO+Debian. We put a lot of effort to keep all distros in sync in regards to the features and capabilities we provide.

Patch-set: 1
Label: Code-Review=-2