Update patch set 10
Patch Set 10:

hi everyone, I think the content in this patch for os-nosdn-ovs scenario is perfect for enabling vxlan provider network across tenant VMs.
As I mentioned earlier, we can enable flat network mappings for external network connectivity once we have required fixes from neutron role (or) it can be configured manually once xci installation is complete.
It would be great if you review this patch again and merge it. This enables me to work on integrating this scenario with releng-xci.

Patch-set: 10