Update patch set 10
Patch Set 10:

> OK then. We will have to wait until the CI is green again for this
 > to go in.
 > >> Do you think we need to really wait for CI to be green ?
 > Because this scenario is nowhere executed unless we execute it from
 > releng-xci. I think we can exempt CI to check-in this patch. This
 > would enable us to test its dependent patch.

Gerrit will not allow us to merge stuff until CI gives +1. There are ways to override that but there is no point in testing the other patch right now since CI will fail. The reason the CI fails is an upstream issue and we are waiting for upstream to merge a fix.

Patch-set: 10