Update patch set 2
Patch Set 2:

To be honest, the better would be to push that in a scenario, as a VIM on baremetal or a VIM inside Openstack. But how to handle it in a scenario, as it can be handle in many scenarios. Maybe the better would be to push that in github or other. I will have the same problem with the designate+bind9 roles i want to add (that is a requirement for ONAP)

Patch-set: 2
Reviewer: Gerrit User 361 <361@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0