Update patch set 2
Patch Set 2:

> I am sorry but I don't like this to be honest. It's far too
 > invasive. You can configure a proxy on per host basis much easier
 > using group variables which should remove 99% of your changes.
 > 
 > http://docs.ansible.com/ansible/latest/playbooks_environment.html

Agree.  I tried to reduce the impact of this change putting as much as I could into a role.  Something that I noticed was that you have to configure the proxy before the usage of user_variables, kind of chicken & egg problem.  I followed this post:

* https://blog.christophersmart.com/2016/08/09/setting-up-openstack-ansible-all-in-one-behind-a-proxy/

Patch-set: 2
Label: Verified=0