Update patch set 2
Patch Set 2:

> >> playbook_dir is an Ansible variable so I believe that the
 > current code is correct. do you have any  problems with it?
 > 
 > Actually we run the ${OPNFV_XCI_PATH}/playbooks/configure-opnfvhost.yml
 > from /tmp/.xci-deploy-env/releng-xci/... path in xci-deploy.sh, but
 > we clone scenarios into XCI_PATH. That's why the following error
 > "<epalper> ERROR! the role 'os-nosdn-ovs' was not found in
 > /tmp/.xci-deploy-env/releng-xci/xci/playbooks/roles:/etc/ansible/roles:/tmp/.xci-deploy-env/releng-xci/xci/playbooks
 > <epalper> This happens with (or) without OPNFV_RELENG_DEV_PATH set"
 > is seen.
 > 
 > Now testing it with following changes:
 > 
 > scenario_path_default: "{{ lookup('env', 'OPNFV_RELENG_PATH')
 > }}/xci/.cache/repos/scenarios"
 > role_path_default: "{{ lookup('env', 'OPNFV_RELENG_PATH')
 > }}/xci/playbooks/roles"

OK but I think the lookup is not needed there. These are exported as ansible variables so you can simply use {{ OPNFV_RELENG_PATH }}. Moreover, it does not matter where we clone the scenarios really because in the end we symlink them to where we want them. playbook_dir should point to "/tmp/.xci-deploy-env/releng-xci/xci/playbooks/" right? That should be the correct location.

Patch-set: 2