Update patch set 3
Patch Set 3:

> Moreover, it does not matter where we clone the scenarios really
 > because in the end we symlink them to where we want them.
 > playbook_dir should point to "/tmp/.xci-deploy-env/releng-xci/xci/playbooks/"
 > right? That should be the correct location.
 > 
 > >> Yes, you are correct. scenarios clone location can be anywhere.
 > Let me change only role_path_default path.

But why do you want to change that? What's wrong with using {{ playbook_dir }}?

 > Now executing get-opnfv-scenario-requirements.yml only after
 > configure-localhost.yml. I think this is appropriate because
 > bootstrap-scenarios happens in configure-opnfvhost.yml.

I am sorry I do not understand what the problem really is. You need the scenarios to be present on localhost. It doesn't matter much if you do it before or after the configure-localhost playbook. What are you trying to fix? What is the problem you are facing?

Patch-set: 3