xci: playbooks: provision-vm-nodes: Simplify the provision VM playbook
The provision-vm-nodes playbook is only responsible for running bifrost
to provision the XCI virtual machines. As such, drop all the extra tasks
that this playbook performed in order to simplify it. This also drops
the XCI_EXTRA_VARS_PATH variable since it had a rather obscure usage
and there is not much value in passing inventory information outside of
XCI. All these will be handled in the scenarios themselves.

Change-Id: If5cb381a3d1e101100eb04478d80fb3045cdfaf8
Signed-off-by: Markos Chandras <mchandras@suse.de>