xci: Make it easier to use OpenStack-Ansible stable branches
It seems it is desirable for projects to use stable OpenStak-Ansible and
OpenStack releases in general. However, merely overriding
OPENSTACK_OSA_VERSION doesn't work because role requirements and
OpenStack services have to be adapted too. Despite the Documentation
warning users about this, seems like nobody remembers to read and
do the necessary steps. However, this leads to numerous invalid bug
reports due to this stable/master mix. Lets try to help improving
this situation by doing all the necessary ground work if we detect that
we are using a stable branch.

Change-Id: I2ba5d7e7942109973b6a6523aae2c01b65e61fe7
Signed-off-by: Markos Chandras <mchandras@suse.de>