bifrost: destroy-env.sh: Clean dib images by default
The destroy-env.sh script is meant to clean the entire environment
including the dib images that were built in a previous run. However,
this step was only executed if requested exclicitly leaving old
artifacts behind. As such, lets switch the default value in order to
remove these images as expected.

Change-Id: If3b91a621df50931ef3d666737cbc3bbea5d34b9
Signed-off-by: Markos Chandras <mchandras@suse.de>