stabilization - structure
This patch provide the main structure for the infra manager that will deploy
the needed servers to receive XCI VIM.

All the deployments options are based on OPNFV PDF/IDF config files and as less
as possible environment variables.

For now the code is out of the main XCI folder, to allow testing and avoid
unwanted pertubations. This will progressivly be move to the main structure.

Change-Id: Ia759497d16743b0d685f3013c15c30a61038552b
Signed-off-by: Blaisonneau David <david.blaisonneau@orange.com>