Update patch set 2
Patch Set 2: Code-Review-1

I am not sure I agree wit this especially when it's enabled by default and without an option to turn it off. Why would we want to have a profiler enabled for every deployment since it will slow it down.

If you are looking to profile the deployment wouldn't it make more sense to enable that in ~/.ansible.cfg for yourself?

If you still want to have this globally enabled, please do not turn it on by default.

Patch-set: 2
Label: Code-Review=-1