bifrost: Set vCPU model to 'host-passthrough'
The default CPU model of QEMU GenuineIntel/QEMU Virtual CPU is not the
best option for vCPU in terms of performance. As such, lets use the
real host CPU. This has also been submitted upstream and the template
should be removed once the upstream patch has been merged and the SHA
for bifrost is bumped.

Change-Id: I23d86933604c290d5650b73f420972d0a2b23dc8
Link: https://review.openstack.org/#/c/528677/
Signed-off-by: Markos Chandras <mchandras@suse.de>