xci: playbooks: Copy local changes to scenarios
Make sure that any local changes to scenarios are being propagated to
the checkouts. This should only happen when we are testing the 'master'
branch of the scenario since it may be desirable to test a specific SHA
in which case we shouldn't modify the checkout code. This also fixes
a typo when creating the scenario/role directories.

Change-Id: I4197f09a3e4d5a54c86905bc556b06e08948fbc2
Signed-off-by: Markos Chandras <mchandras@suse.de>