xci: scenarios: Fix location when copying scenarios to roles
The location where we were putting the scenarios was wrong as it was
placing them in xci/playbooks/roles/$scenario/$scenario so in the end
the Ansible couldn't find the role we were looking for. As a result,
the role was never executed. Moreover, we can use the synchronize
module instead of the shell which should be both faster and safer.

Change-Id: I3d996652bb6a62d23e259da5674d94596dad4d8d
Signed-off-by: Markos Chandras <mchandras@suse.de>