Use explicit path with with_first_found
On the OPNFV gerrit, running on CentOS triggers an Ansible issue that is
similar to https://github.com/ansible/ansible/issues/23392:
the search path for include_vars using with_first_found does not appear
to include the vars directory in the role. This patch is an attempt to
work around that by using the method described above.

For an error that happens without this patch, see
https://build.opnfv.org/ci/job/xci-verify-centos-deploy-virtual-master/503/console

Change-Id: Ie9d421deb1d872e4231f6d6c46dbd5ea1718cd3a
Signed-off-by: Tapio Tallgren <tapio.tallgren@nokia.com>