Update patch set 2
Patch Set 2:

> > >> Can't we merge this with https://gerrit.opnfv.org/gerrit/#/c/49261/
 > >
 > > This is scenario specific file getting used by /c/49261/ in aio
 > > flavored deployment. so we should merge this first to get CI to
 > > pass on the other review.
 > 
 > So why not merge this to the other patchset then?

What I am trying to say is that this patchset can't be tested in the CI on its own. You need the rest of the AIO support from the other patchset so either you need to submit both of them as a patchset series (ie pathces belonging to the same topic) or merge the two patchsets in a single one.

Patch-set: 2