Reduce the IP range
The IP range used to deploy OpenStack services is wide (CIDR /22
offers 1022 Usable IPs). As consequence, it's hard to determine
which IPs are going to be assigned. This can be a challenge to
manage, specially when it's necessary to register them in a no_proxy
environment value. This change reduces the IP range to CIDR /25.

Change-Id: Id3d93b8cc8c20728525374e634098fb98fcf341c
Signed-off-by: Victor Morales <victor.morales@intel.com>