Update patch set 5
Patch Set 5:

So this does 2 things. It adds a role, plus using env variables on all existing roles. Is there no existing role from galaxy to use for setting up a proxy per distro? For the second item, in the past I have suggested to use a single ansible.cfg file to pass your proxy information. Why is that no an option? And it also needs to be split into 2 patches since you are introducing two changes.

Patch-set: 5