Update patch set 4
Patch Set 4:

Why is this now a problem? What changed with these SHAs that made eth12 disappear on ubuntu?
We previously don't create eth12 interface on the controller node which was not required anyway. 
But now neutron agent functionality is moved into controller node and this requires eth12 interface for vlan provider network connectivity.

Patch-set: 4