Pass the OSA version to functest container
OPNFV scenarios are forced to support at least two versions of OSA. If there
are API changes between two supported versions of OpenStack, the testing code
must have two versions. However, without knowing what OpenStack version is
deployed, it is impossible to automate the decision of which testing code to
use. We could pass that information as a env variable to the functest container

Change-Id: I766ea71b576f674f6d488dc9df8effe614e85b28
Signed-off-by: Manuel Buil <mbuil@suse.com>