Add subproject INFO.yaml
This change adds INFO.yaml for releng-xci Gerrit Project which is 
constructed by subset of committers of Releng project who are active in
xci. New committers to releng-xci Gerrit Project will be nominated
on Gerrit via separate changes and voted by releng-xci committers that
are listed within releng-xci/INFO.yaml file as of today.

A single mail with all the nominations will be sent to opnfv-tsc and 
opnfv-tech-discuss mailing lists announcing nominations. A followup
mail will also be sent to announce the results.

Change-Id: I1d49585276304b29ac86b03aff5fd45974a02ccc
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>