Update patch set 1
Patch Set 1:

No I think you misunderstood me. What I meant was which component has this bug? It's not Ansible because osa works fine with dashes. And if it's a bug somewhere in kolla why not fixing kolla instead of using a workaround? I am not familiar with kolla so maybe my questions do not make sense.

Patch-set: 1
Label: Verified=0