Pin kubespray to working sha
Kubespray works fine on supported distros so it is important we pin
the sha and bump it in a controlled manner since people might be trying
xci/k8 scenario and we want them to have working versions.

Apart from that, scenario specific verify/merge jobs are being created
at the moment and we need a working version to test the CI/testing side
of things.

Change-Id: I8702ed697057386449dc99d8cd14391de18f345a
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>