Configure nodes network using PDF/IDF
Add a role to setup the network on the deployed nodes using the
PDF+IDF+Flavor. It also prepare host_vars config for each node to
simplify the next step with the VIM installer. This role can set ip
configuration on physical nics (or vlan) or on linux bridge interface.

4 networks are prepared using the following profiles

network_profiles:
  controller:
    - net-mgmt
  storage:
    - net-mgmt
    - net-storage
  network:
    - net-mgmt
    - net-internal
    - net-vnf
  compute:
    - net-mgmt
    - net-internal
    - net-vnf

If the set_bridges is true following bridges are prepared:

bridge_mapping:
  net-mgmt: br-mgmt
  net-internal: br-vxlan
  net-vnf: br-vlan
  net-storage: br-storage

Change-Id: I269576057650d28100ef5adf72c74a7e474ea682