Update patch set 1
Patch Set 1:

> No I think you misunderstood me. What I meant was which component
 > has this bug? It's not Ansible because osa works fine with dashes.
 > And if it's a bug somewhere in kolla why not fixing kolla instead
 > of using a workaround? I am not familiar with kolla so maybe my
 > questions do not make sense.

Well, I'm also in process to clarify this to me. This is a little context of the problem. Kolla-Ansible uses the globals.yml as configuration file where defines which nics are going to be used, the api_interface is one of them[1]. During the deployment, those nic values are retrieved like this:

    hostvars[inventory_hostname]['ansible_' + api_interface]['ipv4'] 

My understanding is that ansible gathers those custom facts modifying their values, so if the user provides "api_interface: br-vlan" this will be translated to "hostvars[inventory_hostname]['ansible_br_vlan']" instead of "hostvars[inventory_hostname]['ansible_br-vlan']" as consequence the previous error is raised.

The Kolla community has been solve this mismatch including a warning message[2] in their documentation. I don't know the reasons behind but what I found in the bug that I included in the commit message was referring to the nature of ansible.

[1] https://github.com/openstack/kolla-ansible/blob/master/etc/kolla/globals.yml#L90
[2] https://docs.openstack.org/kolla-ansible/latest/admin/production-architecture-guide.html#network-configuration

Patch-set: 1
Label: Verified=0