Connect ODL CI-CD to XCI
FEEDBACK NEEDED: IS THIS DESIGN OK?

Provide the possibility to select what version of ODL to use, master included.
Before executing xci-deploy.sh, the user can specify ODL's version by:

export ODL_VERSION=foo where foo={nitrogen, oxygen, master}

Before bootstrapping the available scenarios, the resolve-odl-repo-url.yml will
generate the repo url pointing to the ODL rpm which matches ODL_VERSION and the
distro which is being used for XCI. That url is the value of the variable
"generated_odl_repo_url" which is passed to the roles that require ODL through
a variable called odl_repo_url.

If the roles would like to have this functionality, they will need to write in
one of their user_* config files the variable odl_repo_url with the url value
generated by XCI. This patch has an example of how to do this with the
os-odl-nofeature scenario.

That variable will be read by os_neutron and passed to the opendaylight role
once these two patches are merged in OSA:

https://review.openstack.org/#/c/550004/
https://review.openstack.org/#/c/549288/

ODL CI-CD pipeline is building one rpm per day in each distro. All those rpms
are published in their nexus.opendaylight.org repo. Therefore, the generated
repos point always to the latest code in nitrogen, oxygen and master(fluorine).
For the time being, if the user wants to deploy with a released version, he/she
must not export any value for ODL_VERSION and the latest released version will
be deployed (exactly what is done today, now pointing to Nitrogen SR2).

Change-Id: I5b3879b0c197059b3aecace48760e983c65a586b
Signed-off-by: Manuel Buil <mbuil@suse.com>