Update patch set 2
Patch Set 2:

> (1 comment)

because we don't clone requirements repo at all so the uc file is not available to us. see the previous patchset - it failed due to not having this env var is not set.

https://build.opnfv.org/ci/job/xci-verify-ubuntu-deploy-virtual-master/877/console

when I checked zuul for bifrost, it seems that repo was cloned for bifrost job. I preferred to go the easy path and set the var instead of messing with the cloning parts which we can easily remove once the problem is resolved. this solution should also be applicable to bifrost-verify jobs we have as well.

Patch-set: 2
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0