xci: scripts: vm: Update package manager cache and ignore errors
We need to make sure that the local package cache is updated before we
try to install packages. Moreover, from time to time, something in the
repos may be broken so we don't want to break the CI job. As such, make
that step non-fatal as simply send the output to stdout so we can see
what's wrong.

Change-Id: I148e3b5435c604bb0d54339935255740c8ce07c0
Signed-off-by: Markos Chandras <mchandras@suse.de>