xci: osa: Switch git repositories to github.com
CI sometimes fails with the following error

"msg": "fatal: unable to access 'https://git.openstack.org/openstack/dragonflow/':
Failed to connect to git.openstack.org port 443: Connection timed out"

We assume that the upstream OpenStack mirrors may get very busy at times
or the connectivity between the CI and the OpenStack repos is not very
stable. As such, lets switch to github.com for the majority of OSA
repositories such as the roles and the services which may improve the
situation.

Change-Id: Ia4668d692077a785c76adeda128eb4cf9f4516a7
Signed-off-by: Markos Chandras <mchandras@suse.de>