Integrate functest-kubernetes into kubespray
This change create a playbook and the role to prepare the environment
for functest-kubernetes execution. The playbook and the role is kept
installer specific even though they can be combined to one. This is a
proposal to separate test environment preparation for different installers
to ease the maintenance at the beginning which can be merged together
if desired.

Change-Id: I6606cfe82e9864118ae5e09c58e81c150ff47763
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>