Update patch set 8
Patch Set 8: Code-Review-1

things are not as simple as just adding new installer into opnfv-scenario-requirements since this will break the ci. look at how things fit together and propose a solution that either continues working as is or ensure changes are done in other places as well and probably before this one.

Patch-set: 8
Label: Code-Review=-1