Update patch set 1
Patch Set 1:

> > (1 comment)
 > >
 > > I'd prefer that the install-lib.sh contains a main function that
 > > installs all the libraries from there instead of calling them in
 > > the xci-deploy.sh, in that way users don't have to navigate
 > between
 > > different files.
 > 
 > I don't understand. xci-deploy.sh doesn't install any other
 > libraries. Could you explain what you mean? What other things
 > should go into that library? This is the first patch of the many. I
 > don't want to dump everything to that file in one go in order to
 > facilitate the reviews

AFAIU more functions are going to be included into install-ansible.sh, if every function is going to be called in different and specific places like this[1], this is the right approach. But if this change pretends to collects all the installations in a single library and only call them in the xci-deploy.sh,  I don't see the reason to call them one by one and maybe a single main function that groups them is fine.

[1] https://gerrit.opnfv.org/gerrit/#/c/54099/1/xci/xci-deploy.sh@109

Patch-set: 1