DNM: Set additional variables for functest
CI_LOOP, NODE_NAME, and BUILD_TAG are needed for logging info to console.
FUNCTEST_MODE and FUNCTEST_SUITE_NAME are important for stating what level
of testing we do for verify and merge jobs.

Setting scenario to os-odl-nofeature as well so we can promote that scenario
as well once this patch gets merged.

deploy-scenario:os-odl-nofeature
installer-type:osa

Change-Id: Ieca09ab6e210b6516a9809f425bb587568fa616e
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>