xci: bifrost: Do not use libvirt as DNS server
Commit 0d332a80cf731e5927c81c9f6929a8b83d43cddd ("Add proxy support")
switched the default DNS server to the libvirt bridge. However, since
we are now configuring the proxy server on the OPNFV host, we do not
have to do that anymore. For deployments behind proxy etc, users can
still use the XCI_ANSIBLE_PARAMS variable to override the default
ipv4_nameserver value.

Change-Id: I7d8fe8c10a1aba2db4a703a81e74ef76fa593d95
Signed-off-by: Markos Chandras <mchandras@suse.de>