Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > Using this approach the OPNFV node is not able to reach any
 > external site
 > 
 > root@opnfv:~# cat /etc/resolv.conf
 > nameserver 8.8.8.8
 > root@opnfv:~# curl http://google.com
 > curl: (6) Could not resolve host: google.com

OK this doesn't work for your setup but is it expected for external DNS to be blocked on such setups? I am not sure whether using the libvirt bridge as DNS is expected to work either. Moreover, we have no global variable to indicate whether a proxy is used so we have to do something like this in the end

ipv4_nameserver: {{ lookup('env', 'http_proxy') | bool | ternary('192.168.122.1', '8.8.8.8') }}

Patch-set: 1
Label: Verified=0